refactor: Rename form to forms
